BTP-20/setup frontend (#7)
* add mantine with a custom theme

* ignore type error from including DOM in tsconfig

* setup auth client

* setup a basic app layout

* setup routing and form validation with sign-in and register pages

* add a guard for pages that should not be seen when a user is signed in

* add an app-link component that applies automatic styling

* add a 404 page

* add a basic marketing page

* add app logos

* split the marketing page into smaller components

* setup component tests and add some

* add a test for the unauthenticated guard

* run code formatter

+++ tests/preload.ts
@@ -1,15 +1,25 @@
-import { beforeEach, afterEach, beforeAll } from "bun:test";
-import { db } from "../app/utils/db";
+import { db } from "$/utils/db";
+import { GlobalRegistrator } from "@happy-dom/global-registrator";
+import { afterEach, beforeAll, beforeEach } from "bun:test";
 
-beforeAll(async () => {
-  await Bun.$`bunx -bun prisma migrate reset --force`;
-  await Bun.$`bun db:seed`;
-});
+GlobalRegistrator.register();
 
-beforeEach(() => {
-  db.$executeRaw`SAVEPOINT before_test`;
-});
+if (!process.env.SKIP_DB_SETUP) {
+  beforeAll(async () => {
+    await Bun.$`bunx -bun prisma migrate reset --force`;
+    await Bun.$`bun db:seed`;
+  });
 
-afterEach(() => {
-  db.$executeRaw`ROLLBACK TO SAVEPOINT before_test`;
+  beforeEach(() => {
+    db.$executeRaw`SAVEPOINT before_test`;
+  });
+
+  afterEach(() => {
+    db.$executeRaw`ROLLBACK TO SAVEPOINT before_test`;
+  });
+}
+
+afterEach(async () => {
+  const { cleanup } = await import("@testing-library/react");
+  cleanup();
 });
